requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
pytrends>=4.9.0
newsapi-python>=0.2.7

# Database
sqlalchemy==1.4.50
psycopg2-binary>=2.9.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
